+++ src/formated_notion_text_example.py
@@ -1,0 +1,83 @@
+{
+  "object": "block",
+  "id": "73e88d7b-bc6d-4125-9364-82ecb71dd26b",
+  "created_time": "2021-12-13T19:56:00.000Z",
+  "last_edited_time": "2022-01-09T16:33:00.000Z",
+  "has_children": True,
+  "archived": False,
+  "type": "to_do",
+  "to_do": {
+    "text": [
+      {
+        "type": "text",
+        "text": {
+          "content": "This is simple text, ",
+          "link": None
+        },
+        "annotations": {
+          "bold": False,
+          "italic": False,
+          "strikethrough": False,
+          "underline": False,
+          "code": False,
+          "color": "default"
+        },
+        "plain_text": "This is simple text, ",
+        "href": None
+      },
+      {
+        "type": "text",
+        "text": {
+          "content": "this is bold text ",
+          "link": None
+        },
+        "annotations": {
+          "bold": True,
+          "italic": False,
+          "strikethrough": False,
+          "underline": False,
+          "code": False,
+          "color": "default"
+        },
+        "plain_text": "this is bold text ",
+        "href": None
+      },
+      {
+        "type": "text",
+        "text": {
+          "content": "and this is ",
+          "link": None
+        },
+        "annotations": {
+          "bold": False,
+          "italic": False,
+          "strikethrough": False,
+          "underline": False,
+          "code": False,
+          "color": "default"
+        },
+        "plain_text": "and this is ",
+        "href": None
+      },
+      {
+        "type": "text",
+        "text": {
+          "content": "a link",
+          "link": {
+            "url": "https://developers.notion.com/reference/rich-text"
+          },
+        "annotations": {
+          "bold": False,
+          "italic": False,
+          "strikethrough": False,
+          "underline": False,
+          "code": False,
+          "color": "default"
+        },
+        "href": "https://developers.notion.com/reference/rich-text"
+        }
+      }
+    ],
+    "checked": False
+  }
+}